requirements.txt
numpy==1.25.2
matplotlib
pillow
scikit-learn
imageio
pandas
seaborn
jupyter
torch        # PyTorch CPU/GPU version
torchvision  # For CNN image transforms if needed



# pip install -r requirements.txt
# python 3.10.11